Move teh base exception to domain rather than data management package
