+++ timetable_solutions/domain/base_exceptions.py
