+++ timetable_solutions/domain/data_management/break_/operations.py
@@ -10,7 +10,7 @@ from django.db import IntegrityError
 
 # Local application imports
 from data import constants, models
-from domain.data_management import base_exceptions
+from domain import base_exceptions
 
 
 class UnableToCreateBreak(base_exceptions.UnableToCreateModelInstance):
