+++ timetable_solutions/domain/data_management/classrooms/operations.py
@@ -9,7 +9,7 @@ from django.db.models import ProtectedError
 
 # Local application imports
 from data import models
-from domain.data_management import base_exceptions
+from domain import base_exceptions
 from domain.data_management.classrooms import queries
