+++ timetable_solutions/domain/data_management/year_groups/operations.py
@@ -7,7 +7,7 @@ from django.db import IntegrityError
 
 # Local application imports
 from data import models
-from domain.data_management import base_exceptions
+from domain import base_exceptions
 from domain.data_management.year_groups import queries
